documented some variables in reservoir.c.
Current status is:

maxframesize = no restrictions

if strict_ISO is defined, we use Gaby's less restrictive
interpretation of the standard.

see TODO version 1.30 for a user who has a hardware player which
cannot handle 320kbps frames if the bit reservoir is used.

I guess the final verdict is tough luck :-)

svn path=/trunk/lame/; revision=2106