Patch from Ti Kan <ti@amb.org>: ---snip--- Below is a patch to make Lame 3.95.1 compile and run on AIX, HP-UX, IRIX, Solaris, UnixWare, SCO Open Server and SunOS 4.1.x.
The change involves the call to snprintf() in libmp3lame/id3tag.c.,
but snprintf() is not available on many of these platforms.  I know
that in general replacing snprintf() with sprintf() removes a layer
of string buffer overflow protection, but in this case it should
be safe because the size of string is bounded.
---snip---

svn path=/trunk/lame/; revision=4262