Make autotools files compatible with automake >= 1.12
Newer autotools versions don't seem to support ansi2knr, which is probably
not a problem, given the last decades of compilers accepting ANSI C (and the
rest of the LAME code being written in ANSI C itself).

This fixes ticket #60, with a patch of Brennan Shacklett, from the Gentoo
community.

svn path=/trunk/lame/; revision=6309